+++ backend/src/controllers/vehicle.controller.ts
@@ -1,10 +1,11 @@
 import { Request, Response } from 'express'
+import validator from 'validator'
+import { Users } from '../models/Users'
 import { Vehicle } from '../models/Vehicle'
-
 export class VehicleController {
 	static async getVehicles(req: Request, res: Response): Promise<void> {
 		try {
-			const vehicles = await Vehicle.findAll()
+			const vehicles = await Vehicle.findAll({include:{model:Users, as: "user"}})
 			res.json(vehicles)
 		} catch (error) {
 			res.status(500).json({ error: 'Internal Server Error' })
@@ -13,9 +14,11 @@ export class VehicleController {
 
 	static async getVehicleById(req: Request, res: Response) {
 		const { id } = req.params
-
+		if (!validator.isUUID(id)) {
+			return res.status(400).json({ message: 'Invalid ID' })
+		}
 		try {
-			const vehicle = await Vehicle.findByPk(id)
+			const vehicle = await Vehicle.findByPk(id,{include:{model:Users, as: "user"}})
 			if (!vehicle) {
 				res.status(404).json({ error: 'Vehicle not found' })
 			}
@@ -28,7 +31,11 @@ export class VehicleController {
 	static async createVehicle(req: Request, res: Response) {
 		const { brand, model, color, year, licensePlate, mileage, userId } =
 			req.body
-		const fields = { brand, model, color, year, licensePlate, mileage }
+
+		if (!validator.isUUID(userId)) {
+			return res.status(400).json({ message: 'Invalid ID' })
+		}
+		const fields = { brand, model, color, year, licensePlate, mileage, userId }
 		const emptyFields = []
 
 		for (const [key, value] of Object.entries(fields)) {
@@ -44,6 +51,8 @@ export class VehicleController {
 		}
 
 		try {
+			const user = await Users.findByPk(userId)
+			if (!user) return res.status(400).json({ message: 'User not found' })
 			const newVehicle = await Vehicle.create({
 				brand,
 				model,
@@ -59,8 +68,11 @@ export class VehicleController {
 		}
 	}
 
-	static async updateVehicle(req: Request, res: Response): Promise<void> {
+	static async updateVehicle(req: Request, res: Response) {
 		const { id } = req.params
+		if (!validator.isUUID(id)) {
+			return res.status(400).json({ message: 'Invalid ID' })
+		}
 		const { brand, model, color, year, licensePlate, mileage, userId } =
 			req.body
 
@@ -85,8 +97,11 @@ export class VehicleController {
 		}
 	}
 
-	static async deleteVehicle(req: Request, res: Response): Promise<void> {
+	static async deleteVehicle(req: Request, res: Response) {
 		const { id } = req.params
+		if (!validator.isUUID(id)) {
+			return res.status(400).json({ message: 'Invalid ID' })
+		}
 
 		try {
 			const deletedVehicleCount = await Vehicle.destroy({ where: { id } })
